refactor(game): extract GameSessionManager out of game_service facade
Move the game-session lifecycle + its mutable tracking state out of the
GameService static facade into a new GameSessionManager collaborator
(services/game/game_session_manager.dart), the single owner of the
launch/session flags, current-game metadata, return/exit callbacks, and
the playtime timer.

Moved verbatim (byte-identical bodies): registerGameLaunch,
_startPlaytimeTimer/_stopPlaytimeTimer/_savePlayTime, endGameSession,
checkPendingGameSession, and the four callback setters, plus all the
session state fields. GameService keeps thin delegators for the frozen
public API (isGameLaunched, isGameLaunchInProgress, begin/clearLaunchPending,
launchedEmulatorExe, checkPendingGameSession, endGameSession, the callback
setters) and a private _registerGameLaunch delegator so the 7 launch call
sites stay unchanged. The straddling orchestration methods
(initializeAndroidGameListener, handleAppResumed) stay on the facade and
read session state through GameSessionManager; onScreenStateChanged (a
public settable field) and the process-check helpers stay on the facade.

Dependency is one-directional (facade -> GameSessionManager). No
public-API change; zero importer churn. Dropped now-unused game_repository
and system_repository imports from the facade. Host 1498 -> ~1340.
